Create ExtensionsGoalState
