+++ tests/protocol/test_hostplugin.py
@@ -153,7 +153,7 @@ class TestHostPlugin(HttpRequestPredicates, AgentTestCase):
         with mock_wire_protocol(DATA_FILE_NO_EXT) as protocol:
             # These tests use mock wire data that don't have any extensions (extension config will be empty).
             # Populate the upload blob and set an initial empty status before returning the protocol.
-            ext_conf = protocol.client._goal_state.ext_conf
+            ext_conf = protocol.client.get_ext_conf()
             ext_conf.status_upload_blob = sas_url
             ext_conf.status_upload_blob_type = page_blob_type
